implementa regras de negócio para fabricante e categoria

+++ src/main/java/net/weg/produtos/service/CategoriaService.java
@@ -6,9 +6,11 @@ import net.weg.produtos.model.dto.IDTO;
 import net.weg.produtos.model.entity.Categoria;
 import net.weg.produtos.repository.CategoriaRepository;
 import org.springframework.beans.BeanUtils;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.NoSuchElementException;
 
 @Service
 @AllArgsConstructor
@@ -16,27 +18,48 @@ public class CategoriaService implements IService<Categoria, Integer> {
 
     CategoriaRepository categoriaRepository;
 
+
+    public static void verificar(Categoria categoria)  {
+        if (categoria.getNome() == null || categoria.getNome() == "") {
+            throw new NullPointerException();
+        }
+    }
+
     @Override
-    public Categoria cadastrar(IDTO obj) {
+    public Categoria cadastrar(IDTO obj)  {
         Categoria categoria = new Categoria();
         BeanUtils.copyProperties(obj, categoria);
+        if (categoriaRepository.existsById(categoria.getId())) {
+            throw new DuplicateKeyException("");
+        }
+        verificar(categoria);
         return categoriaRepository.save(categoria);
     }
 
     @Override
-    public Categoria editar(IDTO obj) {
+    public Categoria editar(IDTO obj)  {
         Categoria categoria = new Categoria();
         BeanUtils.copyProperties(obj, categoria);
+        if (!categoriaRepository.existsById(categoria.getId())) {
+            throw new NoSuchElementException();
+        }
+        verificar(categoria);
         return categoriaRepository.save(categoria);
     }
 
     @Override
     public void deletar(Integer integer) {
+        if (!categoriaRepository.existsById(integer)){
+            throw new NoSuchElementException();
+        }
         categoriaRepository.deleteById(integer);
     }
 
     @Override
-    public Categoria buscarUm(Integer integer) {
+    public Categoria buscarUm(Integer integer)  {
+        if (!categoriaRepository.existsById(integer)){
+            throw new NoSuchElementException();
+        }
         return categoriaRepository.findById(integer).get();
     }
